refactor: 유저 기능 패키지 이동 및 일부 로직 정비
