+++ src/main/java/com/example/demo/controller/user/UserController.java
@@ -1,13 +1,13 @@
-package com.example.demo.controller;
+package com.example.demo.controller.user;
 
-import com.example.demo.dto.UserDataInfo;
-import com.example.demo.dto.UserRequestDto;
+import com.example.demo.dto.user.UserDataInfo;
+import com.example.demo.dto.user.UserRequestDto;
 import com.example.demo.entity.User;
 import com.example.demo.jwt.Jwt;
 import com.example.demo.jwt.JwtProvider;
-import com.example.demo.service.RefreshTokenService;
-import com.example.demo.service.UserService;
-import com.example.demo.repository.RefreshTokenRepository;
+import com.example.demo.service.refreshtoken.RefreshTokenService;
+import com.example.demo.service.user.UserService;
+import com.example.demo.validation.LoginMember;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpHeaders;
@@ -16,7 +16,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -42,7 +41,7 @@ public class UserController {
       return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
 
-    Jwt jwt = jwtProvider.createJwt(user.getEmail(), user.getRole());
+    Jwt jwt = jwtProvider.createJwt(user.getId(), user.getEmail(), user.getRole());
     refreshTokenService.saveRefreshToken(user.getId(), jwt.getRefreshToken());
 
     return ResponseEntity.ok()
@@ -51,9 +50,7 @@ public class UserController {
   }
 
   @GetMapping("/users/me")
-  public ResponseEntity<UserDataInfo> me(HttpServletRequest request) {
-    String email = (String) request.getAttribute("userEmail");
-    User user = userService.findByEmail(email);
+  public ResponseEntity<UserDataInfo> me(@LoginMember User user) {
     return ResponseEntity.ok(new UserDataInfo(user));
   }
