+++ src/main/java/com/example/demo/dto/user/UserDataInfo.java
@@ -1,4 +1,4 @@
-package com.example.demo.dto;
+package com.example.demo.dto.user;
 
 import com.example.demo.entity.User;
