+++ src/main/java/com/example/demo/dto/user/UserResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.demo.dto;
+package com.example.demo.dto.user;
 
 public record UserResponseDto(
     String token,
