+++ src/main/java/com/example/demo/repository/user/UserRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.example.demo.repository;
+package com.example.demo.repository.user;
 
 import com.example.demo.entity.User;
 import java.util.Optional;
